refactor: separated pipeline wrapper component
