+++ packages/developer-portal/src/components/pages/v8/pipeline/pipeline-info-wrapper.tsx
@@ -1,0 +1,38 @@
+import React, { useState } from 'react'
+import { ReapitConnectSession } from '@reapit/connect-session'
+import { PipelineModelInterface } from '@reapit/foundations-ts-definitions'
+import { PipelineDeploymentInfo } from './deployment-info'
+import { useEvent, useChannel } from '@harelpls/use-pusher'
+
+export const PipelineWrapper = ({
+  initialPipeline,
+  connection,
+}: {
+  initialPipeline: PipelineModelInterface
+  connection: ReapitConnectSession
+}) => {
+  const [pipeline, setPipeline] = useState<PipelineModelInterface>(initialPipeline)
+  const channel = useChannel(`private-${connection.loginIdentity.developerId}`)
+
+  useEvent<{ pipeline: PipelineModelInterface }>(channel, 'pipeline-runner-update', (event) => {
+    if (!event) {
+      return
+    }
+
+    if (!pipeline) {
+      return
+    }
+
+    if (event.pipeline.id !== pipeline.id) {
+      return
+    }
+
+    setPipeline(event.pipeline)
+  })
+
+  return (
+    <>
+      <PipelineDeploymentInfo channel={channel} pipeline={pipeline} />
+    </>
+  )
+}
